Moved session-timeout
