+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/session-timeout/index.tsx
